Implement https://github.com/dcmeglio/hubitat-bond/issues/13.
This change adds support for calling the Bond Preset command for motorized shades and maps a special setPosition value of "50" to the Preset command. Hardcoding a special value of "50" is a simplification of a similar pattern used for the Hubitat Somfy Z-Wave controller where a user configurable special setPosition value can be mapped to the Somfy My Button preset value, see https://github.com/dennypage/hubitat/tree/master/drivers/zrtsii.